Split Number into Number + Operations
